python/os/run/template/src/requirements.txt
requests
inquirer
GitPython